+++ CHANGELOG.md
@@ -5,6 +5,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [4.3.0]
+### Added:
+- Added task in ANSIBLE to remove the first version of the bigsdb-insertion.service (if present on the VM)
+- HtmlAntiviralMutationsTableBuilder and HtmlAntiviralMutationsTableBuilder classes 
+- Integration in BIGSdb of two assays for Influenza: the "Antiviral resistances" and the "Reference selection"
+
+### Changed
+- For Influenza, html tags are used but using the section tags of the original report
+- Changes in the naming of pcr related fields for Influenza
+- Adapt import for get_bigsdb_config_data
+
+### Fixed
+- Creation of temp file in "rejected_isolates.py"
+
 ## [4.2.0]
 ### Added:
 - Ability to support deployment for multiple species databases on the same VM
@@ -22,7 +36,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Reworked SFTPConnection class to be able to be used in with statements for cleaner exiting
 - snp lineage is now stored as eav_text data with two fields (one for the html table and the other to help filtering queries)
 
-## Fixed
+### Fixed
 - Issue with rotation of the logs for bigsdb-insertion-{species}.service
 - Literal accepting silently wrong values
 - missing snp_lineage results due to changes in the "mongo key" of the assay after updating to pipeline 1.3